### 2023-05-17 #5473 @editorialbot
[PRE REVIEW]: Molearn: a Python package streamlining the design of generative models of biomolecular dynamics
**Submitting author:** @degiacom (<a href="http://orcid.org/0000-0003-4672-471X">Matteo Degiacomi</a>)
**Repository:** https://github.com/Degiacomi-Lab/molearn
**Branch with paper.md** (empty if default branch):
**Version:** 2.0.1
