+++ doc/source/conf.py
@@ -18,11 +18,11 @@ sys.path.insert(0, os.path.abspath('../../src/molearn'))
 # -- Project information -----------------------------------------------------
 
 project = 'molearn'
-copyright = '2021, V. K. Ramaswamy, S. C. Musson, M. T. Degiacomi'
-author = 'V. K. Ramaswamy, S. C. Musson, M. T. Degiacomi'
+copyright = '2021, S. C. Musson, M. T. Degiacomi'
+author = ' S. C. Musson, M. T. Degiacomi'
 
 # The full version, including alpha/beta/rc tags
-release = '1.3.0'
+release = '2.0.1'
 
 
 # -- General configuration ---------------------------------------------------
@@ -52,7 +52,7 @@ pygments_style = 'sphinx'
 # The theme to use for HTML and HTML Help pages.  See the documentation for
 # a list of builtin themes.
 #
-html_theme = 'classic'
+html_theme = 'sphinxdoc'
 
 # Add any paths that contain custom static files (such as style sheets) here,
 # relative to this directory. They are copied after the builtin static files,
