+++ doc/source/models.rst
@@ -1,6 +1,17 @@
 
-Models
-------
+FoldingNet Model
+----------------
 
-.. automodule:: molearn.models.foldingnet.AutoEncoder
+Import from `molearn.models.foldingnet`
+
+.. autoclass:: molearn.models.foldingnet.AutoEncoder
+   :members:
+
+
+CNN model
+---------
+
+import from  `molearn.models.CNN_autoencoder`
+
+.. autoclass:: molearn.models.CNN_autoencoder.Autoencoder
    :members:
