+++ doc/.readthedocs.yaml
@@ -1,5 +1,3 @@
-{% set version = "1.1.0" %}
-
 package:
   name: molearn
   version: {{ version }}
