+++ doc/.readthedocs.yaml
@@ -1,6 +1,5 @@
 package:
   name: molearn
-  version: {{ version }}
 
 source:
   path: ../
