+++ .readthedocs.yaml
@@ -1,0 +1,24 @@
+{% set version = "2.0.1" %}
+
+package:
+  name: molearn
+  version: {{ version }}
+
+
+build:
+  os: "ubuntu-22.04"
+  tools:
+    python: "3.10"
+
+
+requirements:
+  build:
+    - python
+    - setuptools
+  run:
+    - python
+    - numpy
+    - pytorch
+    - pandas
+    - scipy
+
