+++ .readthedocs.yaml
@@ -1,9 +1,6 @@
-{% set version = "2.0.1" %}
-
 package:
   name: molearn
-  version: {{ version }}
-
+  version: 2.0.1
 
 build:
   os: "ubuntu-22.04"
