+++ .readthedocs.yaml
@@ -9,7 +9,7 @@ build:
 
 
 python:
-  version: 3.10
+  version: 3.7
   install:
     - requirements: docs/requirements.txt
 
