+++ docs/requirements.txt
@@ -1,3 +1,4 @@
+--find-links https://download.pytorch.org/whl/torch_stable.html
 numpy
 pytorch
 pandas
