+++ docs/requirements.txt
@@ -1,5 +1,4 @@
---find-links https://download.pytorch.org/whl/torch_stable.html
 numpy
-pytorch
+pytorch --find-links https://download.pytorch.org/whl/torch_stable.html
 pandas
 scipy
