+++ docs/requirements.txt
@@ -1,4 +1,3 @@
 numpy
-pytorch --find-links https://download.pytorch.org/whl/torch_stable.html
 pandas
 scipy
