### 2023-05-19 4086c76
conda rather than pip python in docs
