+++ .readthedocs.yaml
@@ -5,9 +5,8 @@ package:
 build:
   os: "ubuntu-22.04"
   tools:
-    python: "3.10"
+    python: "mambaforge-4.10"
 
 conda:
-  environment:
-    docs/requirements.yml
+  environment: docs/requirements.yml
 
