+++ .readthedocs.yaml
@@ -1,3 +1,5 @@
+version: 2
+
 package:
   name: molearn
   version: 2.0.1
