+++ environment.yml
@@ -3,8 +3,8 @@ channels:
   - conda-forge
   - defaults
 dependencies:
-  - python=3.7
+  - python=3.8
   - numpy
   - pandas
   - scipy
-  - pytorch
+  - pytorch-cpu=1.7
