+++ .readthedocs.yaml
@@ -1,9 +1,5 @@
 version: 2
 
-package:
-  name: molearn
-  version: 2.0.1
-
 build:
   os: "ubuntu-22.04"
   tools:
