### 2023-05-19 cab59f9
Update dope_score.py to catch modeller errors because salilab is shooting itself in the foot with their outdated licence key practices
