### 2023-05-20 4abd374
Update paper.bib
added missing DOIs
