+++ paper/paper.bib
@@ -65,6 +65,7 @@ year = {2021}
 @article{paszke2019pytorch,
   title={Pytorch: An imperative style, high-performance deep learning library},
   author={Paszke, Adam and Gross, Sam and Massa, Francisco and Lerer, Adam and Bradbury, James and Chanan, Gregory and Killeen, Trevor and Lin, Zeming and Gimelshein, Natalia and Antiga, Luca and others},
+  doi = {10.48550/arXiv.1912.01703},
   journal={Advances in neural information processing systems},
   volume={32},
   year={2019}
@@ -72,6 +73,7 @@ year = {2021}
 @article{eastman2017openmm,
   title={OpenMM 7: Rapid development of high performance algorithms for molecular dynamics},
   author={Eastman, Peter and Swails, Jason and Chodera, John D and McGibbon, Robert T and Zhao, Yutong and Beauchamp, Kyle A and Wang, Lee-Ping and Simmonett, Andrew C and Harrigan, Matthew P and Stern, Chaya D and others},
+  doi = {10.1371/journal.pcbi.1005659},
   journal={PLoS computational biology},
   volume={13},
   number={7},
@@ -82,6 +84,7 @@ year = {2021}
 @article{shen2006statistical,
   title={Statistical potential for assessment and prediction of protein structures},
   author={Shen, Min-yi and Sali, Andrej},
+  doi = {10.1110/ps.062416606},
   journal={Protein science},
   volume={15},
   number={11},
@@ -89,17 +92,10 @@ year = {2021}
   year={2006},
   publisher={Wiley Online Library}
 }
-@online{plotly, 
-author = {Plotly Technologies Inc.}, 
-title = {Collaborative data science}, 
-publisher = {Plotly Technologies Inc.}, 
-address = {Montreal, QC}, 
-year = {2015},
-url = {https://plot.ly}
-}
 @article{nguyen2018nglview,
   title={NGLview--interactive molecular graphics for Jupyter notebooks},
   author={Nguyen, Hai and Case, David A and Rose, Alexander S},
+  doi = {10.1093/bioinformatics/btx789},
   journal={Bioinformatics},
   volume={34},
   number={7},
@@ -110,6 +106,7 @@ url = {https://plot.ly}
 @article{frassek2021extended,
   title={An extended autoencoder model for reaction coordinate discovery in rare event molecular dynamics datasets},
   author={Frassek, M and Arjun, A and Bolhuis, PG},
+  doi = {10.1063/5.0058639},
   journal={The Journal of Chemical Physics},
   volume={155},
   number={6},
@@ -120,6 +117,7 @@ url = {https://plot.ly}
 @article{noe2019boltzmann,
   title={Boltzmann generators: Sampling equilibrium states of many-body systems with deep learning},
   author={No{\'e}, Frank and Olsson, Simon and K{\"o}hler, Jonas and Wu, Hao},
+  doi = {10.1126/science.aaw1147},
   journal={Science},
   volume={365},
   number={6457},
@@ -130,6 +128,7 @@ url = {https://plot.ly}
 @article{chen2018collective,
   title={Collective variable discovery and enhanced sampling using autoencoders: Innovations in network architecture and error function design},
   author={Chen, Wei and Tan, Aik Rui and Ferguson, Andrew L},
+  doi = {10.1063/1.5023804},
   journal={The Journal of chemical physics},
   volume={149},
   number={7},
@@ -140,6 +139,7 @@ url = {https://plot.ly}
 @article{sidky2020molecular,
   title={Molecular latent space simulators},
   author={Sidky, Hythem and Chen, Wei and Ferguson, Andrew L},
+  doi = {10.1039/d0sc03635h},
   journal={Chemical Science},
[…]
