+++ paper/paper.md
@@ -33,7 +33,7 @@ example data obtained via experiment or molecular simulation.
 
 # Statement of need
 
-Most biological mechanisms directly involve proteins. The specific task each of these biopolymers carries out is linked to their three-dimensional shape, enabling them to bind to designated binding partners such as small molecules, ions, or other biopolymers. Crucially though, biomolecules are flexible and so are continuously jostling and reconfiguring due to Brownian motion. Thus, their function emerges from characteristic conformational dynamics. Characterising the structure and dynamics of biomolecules at the atomic level provides us with a fundamental understanding of the mechanisms underpinning life and is the first step in numerous technological applications. Progress in these areas has been spearheaded by the development of a diverse palette of dedicated experimental techniques. Unfortunately, none is singlehandedly capable of routinely reporting on the full fine structure of biomolecular conformational spaces. As such, our understanding of life at the molecular level is inherently biased by the techniques we adopt to observe it [@Marsh2015]. Molecular dynamics (MD) simulations yield atomistic insights into the conformational landscape of biomolecules, complementing and extending data gathered experimentally. MD simulations produce samples of molecular conformational spaces by iteratively generating new conformers based on an initial, known atomic arrangement and physical models of atomic interactions. While MD enables obtaining key insight into biomolecular function, it is not a silver bullet: an exhaustive sampling of key biological processes such as folding or spontaneous binding with partners usually lay beyond what can be routinely observed.
+Most biological mechanisms directly involve proteins. The specific task each of these biopolymers carries out is linked to their three-dimensional shape, enabling them to bind to designated binding partners such as small molecules, ions, or other biopolymers. Crucially though, biomolecules are flexible and so are continuously jostling and reconfiguring due to Brownian motion. Thus, their function emerges from characteristic conformational dynamics. Characterising the structure and dynamics of biomolecules at the atomic level provides us with a fundamental understanding of the mechanisms underpinning life and is the first step in numerous technological applications. Progress in these areas has been spearheaded by the development of a diverse palette of dedicated experimental techniques. Unfortunately, none is singlehandedly capable of routinely reporting on the full fine structure of biomolecular conformational spaces. As such, our understanding of life at the molecular level is inherently biased by the techniques we adopt to observe it [@Marsh2015]. Molecular dynamics (MD) simulations yield atomistic insights into the conformational landscape of biomolecules, complementing and extending data gathered experimentally. MD simulations produce samples of molecular conformational spaces by iteratively generating new conformers based on an initial, known atomic arrangement and physical models of atomic interactions. While MD enables obtaining key insight into biomolecular function, it is not a silver bullet: exhaustive sampling of processes such as folding or ligand binding usually lay beyond what can be routinely observed.
 
 Generative Neural Networks (GNNs) have been shown to be effective predictors of a protein's 3D structure from its sequence [@Jumper2021; @Baek2021]. Several efforts have also demonstrated that a neural network trained with MD conformers can learn a meaningful dimensionality reduction of the data, usable for reaction coordinate definition [@frassek2021extended; @chen2018collective], or driving conformational space sampling [@noe2019boltzmann; @sidky2020molecular; @mehdi2022accelerating]. In this context, we have previously presented GNNs capable of generating protein conformations based on small pools of examples produced by MD [@Degiacomi2019; @Ramaswamy2021]. The issue is that developing an Machine Learning model to study biomolecular dynamics is a lengthy process. This requires setting up means of transforming conformational space data into tensor data submittable to a model, as well as assessing a model's quality (e.g., in terms of their energy or according to structural descriptors). Here we present `molearn`, a Python framework facilitating the implementation of generative neural networks learning protein conformational spaces from examples obtained via experiments or MD simulations.
 
