### 2023-05-23 #5473 @RMeli
Yes, I am.

**Disclaimer**: @degiacom was the _external_ examiner at my PhD thesis defence.
