### 2023-05-28 407911a
bugfix in GUI
prevent min>max error when loading GUI
