### 2023-05-30 4f63751
alignment in rmsd landscape (analyser). deepcopy rather than use self.mol directly. reorder arguments in trainer.load_checkpoint
