### 2023-06-07 #5523 @editorialbot
[REVIEW]: Molearn: a Python package streamlining the design of generative models of biomolecular dynamics
**Submitting author:** @degiacom (<a href="http://orcid.org/0000-0003-4672-471X">Matteo Degiacomi</a>)
**Repository:** https://github.com/Degiacomi-Lab/molearn
**Branch with paper.md** (empty if default branch):
**Version:** v2.0.3
