### 2023-06-07 Degiacomi-Lab/molearn#6 @RMeli
[JOSS] Installation
Installation using `mamba` on a fresh environment (w/ Python 3.10) fails on macOS (M1) with the following error:

```text
Could not solve for environment specs
The following package could not be installed
└─ molearn   is uninstallable because it requires
   └─ biobox  , which does not exist (perhaps a missing channel).
```

I think the issue stems from the fact that `molearn`'s conda-forge feedstock is `noarch`, while `biobox`'s confa-forge feedstock is `linux-64 | win-64 | osx-64`.

I'd suggest the following:

* Restrict the supported architectures in the conda-forge package to fail earlier and with a better error message
* Clarify the supported architectures in the README and documentation, under the **Installation** section
  * Windows is also not fully supported as far as I can tell (`openmmtorchplugin` limitation), but this information is only available in the FAQ

---

*Note: I'm travelling with only access to macOS (M1). I'll be back to a workstation with a Linux distribution on Monday.*

https://github.com/openjournals/joss-reviews/issues/5523
