+++ paper/paper.md
@@ -26,16 +26,14 @@ bibliography: paper.bib
 
 # Summary
 
-We present `molearn`, a Python package streamlining the implementation of machine 
-learning models dedicated to the generation of protein conformations from
-example data obtained via experiment or molecular simulation.
+We present `molearn`, a Python package streamlining the implementation of machine learning models dedicated to the generation of protein conformations from example data obtained via experiment or molecular simulation.
 
 
 # Statement of need
 
-Most biological mechanisms directly involve proteins. The specific task each of these biopolymers carries out is linked to their three-dimensional shape, enabling them to bind to designated binding partners such as small molecules, ions, or other biopolymers. Crucially though, biomolecules are flexible and so are continuously jostling and reconfiguring due to Brownian motion. Thus, their function emerges from characteristic conformational dynamics. Characterising the structure and dynamics of biomolecules at the atomic level provides us with a fundamental understanding of the mechanisms underpinning life and is the first step in numerous technological applications. Progress in these areas has been spearheaded by the development of a diverse palette of dedicated experimental techniques. Unfortunately, none is singlehandedly capable of routinely reporting on the full fine structure of biomolecular conformational spaces. As such, our understanding of life at the molecular level is inherently biased by the techniques we adopt to observe it [@Marsh2015]. Molecular dynamics (MD) simulations yield atomistic insights into the conformational landscape of biomolecules, complementing and extending data gathered experimentally. MD simulations produce samples of molecular conformational spaces by iteratively generating new conformers based on an initial, known atomic arrangement and physical models of atomic interactions. While MD enables obtaining key insight into biomolecular function, it is not a silver bullet: exhaustive sampling of processes such as folding or ligand binding usually lay beyond what can be routinely observed.
+Most biological mechanisms directly involve proteins. The specific task each of these biopolymers carries out is linked to their three-dimensional shape, enabling them to bind to designated binding partners such as small molecules, ions, or other biopolymers. Crucially though, biomolecules are flexible and so are continuously jostling and reconfiguring due to Brownian motion. Thus, their function emerges from characteristic conformational dynamics. Characterising the structure and dynamics of biomolecules at the atomic level provides us with a fundamental understanding of the mechanisms underpinning life and is the first step in numerous technological applications. Progress in these areas has been spearheaded by the development of a diverse palette of dedicated experimental techniques [@dobson2019]. Unfortunately, none is singlehandedly capable of routinely reporting on the full fine structure of biomolecular conformational spaces. As such, our understanding of life at the molecular level is inherently biased by the techniques we adopt to observe it [@Marsh2015]. Molecular dynamics (MD) simulations yield atomistic insights into the conformational landscape of biomolecules, complementing and extending data gathered experimentally. MD simulations produce samples of molecular conformational spaces by iteratively generating new conformers based on an initial, known atomic arrangement and physical models of atomic interactions. While MD enables obtaining key insight into biomolecular function, it is not a silver bullet: exhaustive sampling of processes such as folding or ligand binding usually lay beyond what can be routinely observed.
 
[…]
