+++ paper/paper.bib
@@ -11,6 +11,15 @@ title = {{Structure, Dynamics, Assembly, and Evolution of Protein Complexes}},
 volume = {84},
 year = {2015}
 }
+@article{dobson2019,
+  title={Biophysical techniques in structural biology},
+  author={Dobson, Christopher M},
+  journal={Annual review of biochemistry},
+  volume={88},
+  pages={25--33},
+  year={2019},
+  publisher={Annual Reviews}
+}
 @article{Jumper2021,
 abstract = {{\textless}p{\textgreater} Proteins are essential to life, and understanding their structure can facilitate a mechanistic understanding of their function. Through an enormous experimental effort {\textless}sup{\textgreater}1–4{\textless}/sup{\textgreater} , the structures of around 100,000 unique proteins have been determined {\textless}sup{\textgreater}5{\textless}/sup{\textgreater} , but this represents a small fraction of the billions of known protein sequences {\textless}sup{\textgreater}6,7{\textless}/sup{\textgreater} . Structural coverage is bottlenecked by the months to years of painstaking effort required to determine a single protein structure. Accurate computational approaches are needed to address this gap and to enable large-scale structural bioinformatics. Predicting the three-dimensional structure that a protein will adopt based solely on its amino acid sequence—the structure prediction component of the ‘protein folding problem' {\textless}sup{\textgreater}8{\textless}/sup{\textgreater} —has been an important open research problem for more than 50 years {\textless}sup{\textgreater}9{\textless}/sup{\textgreater} . Despite recent progress {\textless}sup{\textgreater}10–14{\textless}/sup{\textgreater} , existing methods fall far short of atomic accuracy, especially when no homologous structure is available. Here we provide the first computational method that can regularly predict protein structures with atomic accuracy even in cases in which no similar structure is known. We validated an entirely redesigned version of our neural network-based model, AlphaFold, in the challenging 14th Critical Assessment of protein Structure Prediction (CASP14) {\textless}sup{\textgreater}15{\textless}/sup{\textgreater} , demonstrating accuracy competitive with experimental structures in a majority of cases and greatly outperforming other methods. Underpinning the latest version of AlphaFold is a novel machine learning approach that incorporates physical and biological knowledge about protein structure, leveraging multi-sequence alignments, into the design of the deep learning algorithm. {\textless}/p{\textgreater}},
 author = {Jumper, John and Evans, Richard and Pritzel, Alexander and Green, Tim and Figurnov, Michael and Ronneberger, Olaf and Tunyasuvunakool, Kathryn and Bates, Russ and {\v{Z}}{\'{i}}dek, Augustin and Potapenko, Anna and Bridgland, Alex and Meyer, Clemens and Kohl, Simon A. A. and Ballard, Andrew J. and Cowie, Andrew and Romera-Paredes, Bernardino and Nikolov, Stanislav and Jain, Rishub and Adler, Jonas and Back, Trevor and Petersen, Stig and Reiman, David and Clancy, Ellen and Zielinski, Michal and Steinegger, Martin and Pacholska, Michalina and Berghammer, Tamas and Bodenstein, Sebastian and Silver, David and Vinyals, Oriol and Senior, Andrew W. and Kavukcuoglu, Koray and Kohli, Pushmeet and Hassabis, Demis},
