+++ paper/paper.bib
@@ -14,6 +14,7 @@ year = {2015}
 @article{dobson2019,
   title={Biophysical techniques in structural biology},
   author={Dobson, Christopher M},
+  doi = {10.1146/annurev-biochem-013118-111947},
   journal={Annual review of biochemistry},
   volume={88},
   pages={25--33},
