### 2023-06-10 4aa4b8f
Update paper.bib
added refs to MDAnalysis, biobox, and numpy
