+++ paper/paper.bib
@@ -179,3 +179,36 @@ year = {2021}
   year={2015},
   publisher={ACS Publications}
 }
+@article{rudden2022biobox,
+  title={Biobox: a toolbox for biomolecular modelling},
+  author={Rudden, Lucas SP and Musson, Samuel C and Benesch, Justin LP and Degiacomi, Matteo T},
+  doi={10.1093/bioinformatics/btab785},
+  journal={Bioinformatics},
+  volume={38},
+  number={4},
+  pages={1149},
+  year={2022},
+  publisher={Oxford University Press}
+}
+@article{michaud2011mdanalysis,
+  title={MDAnalysis: a toolkit for the analysis of molecular dynamics simulations},
+  author={Michaud-Agrawal, Naveen and Denning, Elizabeth J and Woolf, Thomas B and Beckstein, Oliver},
+  doi={10.1002/jcc.21787}
+  journal={Journal of computational chemistry},
+  volume={32},
+  number={10},
+  pages={2319--2327},
+  year={2011},
+  publisher={Wiley Online Library}
+}
+@article{harris2020array,
+  title={Array programming with NumPy},
+  author={Harris, Charles R and Millman, K Jarrod and Van Der Walt, St{\'e}fan J and Gommers, Ralf and Virtanen, Pauli and Cournapeau, David and Wieser, Eric and Taylor, Julian and Berg, Sebastian and Smith, Nathaniel J and others},
+  doi={10.1038/s41586-020-2649-2}
+  journal={Nature},
+  volume={585},
+  number={7825},
+  pages={357--362},
+  year={2020},
+  publisher={Nature Publishing Group UK London}
+}
