### 2023-06-10 620da4a
Update paper.md
added explicit mention to MDAnalysis, biobox, and numpy
