+++ paper/paper.bib
@@ -193,7 +193,7 @@ year = {2021}
 @article{michaud2011mdanalysis,
   title={MDAnalysis: a toolkit for the analysis of molecular dynamics simulations},
   author={Michaud-Agrawal, Naveen and Denning, Elizabeth J and Woolf, Thomas B and Beckstein, Oliver},
-  doi={10.1002/jcc.21787}
+  doi={10.1002/jcc.21787},
   journal={Journal of computational chemistry},
   volume={32},
   number={10},
@@ -204,7 +204,7 @@ year = {2021}
 @article{harris2020array,
   title={Array programming with NumPy},
   author={Harris, Charles R and Millman, K Jarrod and Van Der Walt, St{\'e}fan J and Gommers, Ralf and Virtanen, Pauli and Cournapeau, David and Wieser, Eric and Taylor, Julian and Berg, Sebastian and Smith, Nathaniel J and others},
-  doi={10.1038/s41586-020-2649-2}
+  doi={10.1038/s41586-020-2649-2},
   journal={Nature},
   volume={585},
   number={7825},
