### 2023-06-10 Degiacomi-Lab/molearn#7 @degiacom
Thank you for your feedback, here is a summary of the changes made to the paper and `README.md`:

* We agree that the GNN acronym can be confusing, so we replaced it with everywhere in the text.
* We applied the modifications to the text as suggested
* In the Introduction, we added a reference to the 2019 Annual Reviews opening article by Chris Dobson, discussing experimental techniques developed to study protein structure, as well as the importance of protein dynamics in determining their function.
* Concerning the dependencies described in the `README.md` file, we have now added references to Biobox and MDAnalysis (used for I/O) and numpy (our analysis part of the package returns arrays). geomloss is associated with an experimental feature we only recently started investigating, we would rather not highlight this feature at this stage.
* When discussing "Model Analysis" (lines 58-61), we realised the we used the word "model" referring to both the neural network, and the protein structures it generates. We have rephrased two sentences to disambiguate these two concepts.
* We now mention NGLView in the `README.md` file, and link to its webpage.
