### 2023-06-10 e16c001
Update README.md
Mention that Windows is not supported for energy evaluation via OpenMM
