### 2023-06-10 Degiacomi-Lab/molearn#6 @degiacom
Your observation that energy evaluation via OpenMM is not supported on Windows is correct, this is now also stated in the README. Thank you also for mentioning a problem with installation on macOS. Overall, I think the best solution would be to restrict the architecture to Linux only (@SCMusson is looking into this).
