### 2023-06-10 #5523 @degiacom
@RMeli, thank you for your feedback! We have updated the paper according to your suggestions (see https://github.com/Degiacomi-Lab/molearn/issues/7#issuecomment-1585712845), and are looking into the installation issue on MacOS you reported.
