### 2023-06-11 Degiacomi-Lab/molearn#6 @RMeli
> Overall, I think the best solution would be to restrict the architecture to Linux only

Sounds very sensible to me. `mamba install molearn` on Ubuntu 23.04 seems to work as expected. Probably a simple note on the `README.md` is enough, but if you want to go the extra mile and restrict it on the conda-forge package too I think it would be great. You can probably use [preprocessing selectors](https://docs.conda.io/projects/conda-build/en/stable/resources/define-metadata.html#preprocessing-selectors)?

```
build:
   skip: True  # [win or osx]
```

---

Is there any reason why `setup.py` does not contain the dependencies under `install_requires` and the optional dependencies under the `extra_requires` fields? If you decide to restrict the conda-forge package, it might be worth adding `platforms = ["Linux"]` to `setup.py` too.
