### 2023-06-11 Degiacomi-Lab/molearn#6 @RMeli
Under the "Installation" section in the `README.md` and the documentation, I'd mention that `modeller` requires an academic license. I'd also mention which dependencies are automatically installed by `conda install molearn -c conda-forge` (the required ones, plus `openmmtorchplugin`) and state that most optional dependencies need to be manually installed.
