### 2023-06-11 Degiacomi-Lab/molearn#6 @RMeli
`plotly` seems to be a dependency of the (optional) GUI too?

https://github.com/Degiacomi-Lab/molearn/blob/e16c0018e096a3903522941c8a72aa1defb9cead/src/molearn/analysis/GUI.py#L27
