### 2023-06-12 Degiacomi-Lab/molearn#6 @degiacom
I added a reference to plotly, mentioned that modeller requires academic license, and indicated that currently only Linux is supported. @SCMusson, while we could sort issues with dependencies with extra text in the README, I think it would make sense to update the conda installation, what do you think?
