### 2023-06-12 Degiacomi-Lab/molearn#6 @SCMusson
I agree, I have removed the `noarch` from the molearn-feedstock and now skip builds for `[not linux]`
