### 2023-06-13 Degiacomi-Lab/molearn#6 @RMeli
> while we could sort issues with dependencies with extra text in the README, I think it would make sense to update the conda installation

I was suggesting the path of least resistance, but providing a full (with/without `modeller`?) `conda`/`pip` installation will definitely make things easier for the end user.
