+++ docs/source/scoring.rst
@@ -2,14 +2,14 @@
 Scoring
 -------
 
-.. automodule:: molearn.scoring.DOPE_Score
+.. automclass:: molearn.scoring.DOPE_Score
    :members:
 
-.. automodule:: molearn.scoring.Parallel_DOPE_Score
+.. autoclass:: molearn.scoring.Parallel_DOPE_Score
    :members:
    
-.. automodule:: molearn.scoring.Ramachandran_Score
+.. autoclass:: molearn.scoring.Ramachandran_Score
    :members:
    
-.. automodule:: molearn.scoring.Parallel_Ramachandran_Score
+.. autoclass:: molearn.scoring.Parallel_Ramachandran_Score
    :members:
