### 2023-06-15 73fe392
update readthedocs.yaml
added plotly and nglview to requirements. Also, small formatting cleanup inscoring doc.
