### 2023-06-15 1c96c70
added channels to .readthedocs.yaml
attempt to get reathedocs to compile the analysis page
