+++ docs/.readthedocs.yaml
@@ -10,6 +10,11 @@ build:
 sphinx:
   configuration: docs/source/conf.py
   
+channels:
+  - conda-forge
+  - plotly
+  - defaults
+  
 requirements:
   build:
     - python
