### 2023-06-15 70cc8b6
edit environment.yml
added nglview and plotly dependencies, and plotly channel
