+++ docs/.readthedocs.yaml
@@ -1,33 +1,0 @@
-package:
-  name: molearn
-
-source:
-  path: ../
-
-build:
-  number: 0
-
-sphinx:
-  configuration: docs/source/conf.py
-  
-channels:
-  - conda-forge
-  - plotly
-  - defaults
-  
-requirements:
-  build:
-    - python
-    - setuptools
-  run:
-    - python
-    - numpy
-    - pytorch
-    - pandas
-    - scipy
-    - plotly
-    - nglview
-test:
-  imports:
-    - molearn
-
