### 2023-06-15 070520c
update doc of loss_functions
