+++ src/molearn/loss_functions/openmm_thread.py
@@ -1,8 +1,8 @@
 import os
-#from openmm.unit import kelvin, picosecond
+
 try:
     from openmm import Platform
-    from openmm.app import ForceField, PDBFile, Simulation #, OBC2
+    from openmm.app import ForceField, PDBFile, Simulation
     from openmm.app import element as elem
     import openmm
     from openmm.app.forcefield import _createResidueSignature
@@ -14,11 +14,16 @@ except ImportError as e:
     
 import torch
 import numpy as np
-#from math import ceil
 
 
 class ModifiedForceField(ForceField):
+    
     def __init__(self, *args, alternative_residue_names = None, **kwargs):
+        '''
+        Takes all `*args` and `**kwargs` of `openmm.app.ForceField`, plus an optional parameter described here.
+        
+        :param dict  alternative_residue_names: aliases for resnames, e.g., `{'HIS':'HIE'}`.
+        '''
         super().__init__(*args, **kwargs)
         if isinstance(alternative_residue_names, dict):
             self._alternative_residue_names = alternative_residue_names
@@ -26,23 +31,15 @@ class ModifiedForceField(ForceField):
             self._alternative_residue_names = {'HIS':'HIE'}
 
     def _getResidueTemplateMatches(self, res, bondedToAtom, templateSignatures=None, ignoreExternalBonds=False, ignoreExtraParticles=False):
-        """Return the templates that match a residue, or None if none are found.
+        """
+        Return the templates that match a residue, or None if none are found.
 
-        Parameters
-        ----------
-        res : Topology.Residue
-            The residue for which template matches are to be retrieved.
-        bondedToAtom : list of set of int
-            bondedToAtom[i] is the set of atoms bonded to atom index i
-
-        Returns
-        -------
-        template : _TemplateData
-            The matching forcefield residue template, or None if no matches are found.
-        matches : list
-            a list specifying which atom of the template each atom of the residue
-            corresponds to, or None if it does not match the template
-
+        :param res: Topology.Residue, the residue for which template matches are to be retrieved.
+        :param bondedToAtom: list of set of int, bondedToAtom[i] is the set of atoms bonded to atom index i
+        :returns: list with two elements [template, matches]. 
+                  _TemplateData is the matching forcefield residue template, or None if no matches are found.
+                  matches is a list specifying which atom of the template each atom of the residue corresponds to,
+                  or None if it does not match the template.
         """
         template = None
         matches = None
@@ -82,7 +79,7 @@ class ModifiedForceField(ForceField):
                         template = t
                         matches = m
                         return [template, matches]
-                print(f'multpile for {t.name}')
+                print(f'multiple for {t.name}')
                 # We found multiple matches.  This is OK if and only if they assign identical types and parameters to all atoms.
                 t1, m1 = allMatches[0]
 
@@ -95,20 +92,21 @@ class ModifiedForceField(ForceField):
 
 class OpenmmPluginScore():
     '''
-    This will use the new Openmm Plugin to calculate forces and energy. The intention is that this will be fast enough to be able to calculate forces and energy during training.
-    NB. The current torchintegratorplugin only supports float on GPU and double on CPU.
+    This will use the new OpenMM Plugin to calculate forces and energy. The intention is that this will be fast enough to be able to calculate forces and energy during training.
+    N.B.: The current torchintegratorplugin only supports float on GPU and double on CPU.
     '''
     
[…]
