### 2023-06-23 Degiacomi-Lab/molearn#8 @RMeli
[JOSS] Examples
Running `analysis_example.py` in the `examples/` folder results in the following error:

```python
FileNotFoundError: [Errno 2] No such file or directory: 'xbb_foldingnet_checkpoints\\checkpoint_no_optimizer_state_dict_epoch167_loss0.003259085263643.ckpt'
```

This happens similarly for the following paths:
https://github.com/Degiacomi-Lab/molearn/blob/070520c52945427efef070b28c55805330c98f92/examples/analysis_example.py#L32-L33

I'd suggest to use `os.path.join` to concatenate paths.

---

https://github.com/openjournals/joss-reviews/issues/5523
