### 2023-06-23 Degiacomi-Lab/molearn#8 @RMeli
This is a bonus and not required at all for JOSS, but would it be possible to add a plotting output of the analysis script?

https://github.com/Degiacomi-Lab/molearn/blob/070520c52945427efef070b28c55805330c98f92/examples/analysis_example.py#L50-L58

The "generating error landscape" phase takes some time so the suspense runs high and it's a bit deceiving not to get any output. =)
