### 2023-06-23 Degiacomi-Lab/molearn#8 @RMeli
When running `bb_example_subclassing_trainer.py` I get the following error:

```
Traceback (most recent call last):
  File "/home/rmeli/git/molearn/examples/bb_example_subclassing_trainer.py", line 127, in <module>
    trainer.run(max_epochs = 32+trainer.epoch,**runkwargs)
  File "/home/rmeli/git/molearn/src/molearn/trainers/trainer.py", line 181, in run
    logs.update(self.valid_epoch(epoch))
                ^^^^^^^^^^^^^^^^^^^^^^^
  File "/home/rmeli/git/molearn/examples/bb_example_subclassing_trainer.py", line 87, in valid_epoch
    results['valid_DOPE'] = dope[:,0].mean()
                            ~~~~^^^^^
IndexError: too many indices for array: array is 1-dimensional, but 2 were indexed
```

Printing the shape of `dope` I get `(8,)`.
