### 2023-06-28 68f788c
refine = 'both' to retain old behaviour for dope scores
