### 2023-06-28 #5523 @SCMusson
@RMeli I have added a [`CONTRIBUTING.md`](https://github.com/Degiacomi-Lab/molearn/blob/master/CONTRIBUTING.md) file. Hopefully it satisfies the requirements
