### 2023-06-29 Degiacomi-Lab/molearn#9 @RMeli
[JOSS] Tests
The only test in the `test` folder fails with the following error:

```
Warning: importing 'simtk.openmm' is deprecated.  Import 'openmm' instead.
Dataset.shape: torch.Size([16, 3, 2145])
mean: 44.69796567599069, std: 13.886234009167817
E
ERROR: test_openmm_energy (test_openmm_plugin.Test_openmm_plugin.test_openmm_energy)
Traceback (most recent call last):
  File "/home/rmeli/git/molearn/test/test_openmm_plugin.py", line 30, in test_openmm_energy
    openmmscore = molearn.openmm_energy(data.mol, data.std, clamp=None, platform = 'CUDA') #xml_file = ['modified_amber_protein.xml',])
                  ^^^^^^^^^^^^^^^^^^^^^
AttributeError: module 'molearn' has no attribute 'openmm_energy'

Ran 1 test in 0.062s

FAILED (errors=1)
```

While the functionality of the software can verified through the provided example notebooks and the scripts in the `examples` folder, I think the code could greatly benefit from some additional unit tests (and, eventually, the use of CI).

This relates to my comment in #8; if tests, examples scripts, and notebooks are not run each time the main package is updated they are bound to diverge eventually.

---

Since the `example_script.py` is not testing anything (as far as I can tell), should it be moved elsewhere (for example the `examples` folder)?

---

The only necessary action item here (IMO) is to fix the only test available.

From my understanding of JOSS criteria, building a more extensive test suite and adding CI is not required. Personally I'd strongly recommend it as it will benefit future development, but can be added in the future.

---

https://github.com/openjournals/joss-reviews/issues/5523
