### 2023-07-05 d40322e
added tqdm to analyser + bug fixes
