### 2023-07-05 c0f345c
bugfix for scan_error_from_target
