### 2023-07-06 #5523 @degiacom
@RMeli, thank you for your feedback. Here is a summary of changes:
- we fixed the [problems with Jupyter notebooks](https://github.com/Degiacomi-Lab/molearn_notebook/issues/3#issuecomment-1622685569).
- now analysis also features [progress bars](https://github.com/Degiacomi-Lab/molearn/commit/d40322e8520ccd34f882b1120a610771c099f2b4), to give the user some feedback during calculation of 2D landscapes
- unittesting has been [improved](https://github.com/Degiacomi-Lab/molearn/commit/aac0a6a4b78325125cbd3e9ef1c1aff8495c3ee9)

Throughout, we have also applied  minor feature updates and improved package management during installation. All of these changes will be part of a minor release when the reviewing process completes.
