### 2023-07-06 Degiacomi-Lab/molearn#9 @degiacom
Thank you for your feedback, I agree testing needs to be improved. We have now updated the unittests in [test/test_openmm_plugin.py](https://github.com/Degiacomi-Lab/molearn/blob/master/test/test_openmm_plugin.py). Concerning example_script.py, it is currently used in the CI of conda-forge.
