### 2023-07-07 Degiacomi-Lab/molearn#6 @RMeli
While trying to install `molearn` using `setup.py` to make sure I have the latest version installed, I encountered the following issues:
* `openmmtorchplugin` is not available on PyPI
* PyTorch is called `torch` on PyPI

I changed the  to the following in `setup.py`:
```diff
-    install_requires = ["numpy", "pytorch", "biobox", "openmmtorchplugin"],
+    install_requires = ["numpy", "torch>=1.7", "biobox"],
+    extras_require={
+        "gui": ["MDAnalysis", "plotly", "nglview"],
+        "geomloss": ["geomloss"],
+    }
```
and this works for me
```
pip install -e ".[gui,geomloss]"
```

The OpenMM and DOPE plugins can't be supported by this method, but that's OK as long as this is clear in the documentation. This would allow to easily install a development version of `molearn` (with `pip install -e .`) and the other plugins can be installed/managed with `conda`.
