### 2023-07-08 Degiacomi-Lab/molearn#9 @RMeli
Thanks for updating the test and adding some more! I was able to run them with `python -m coverage run -m unittest` and they all pass.

```
Ran 32 tests in 7.462s

OK
```

```
Name                                                                               Stmts   Miss  Cover
------------------------------------------------------------------------------------------------------
/home/rmeli/git/molearn/src/molearn/__init__.py                                        5      0   100%
/home/rmeli/git/molearn/src/molearn/data/__init__.py                                   2      0   100%
/home/rmeli/git/molearn/src/molearn/data/pdb_data.py                                 122     25    80%
/home/rmeli/git/molearn/src/molearn/loss_functions/__init__.py                         3      0   100%
/home/rmeli/git/molearn/src/molearn/loss_functions/openmm_thread.py                  262    121    54%
/home/rmeli/git/molearn/src/molearn/loss_functions/torch_protein_energy.py           213    191    10%
/home/rmeli/git/molearn/src/molearn/loss_functions/torch_protein_energy_utils.py     761    740     3%
/home/rmeli/git/molearn/src/molearn/trainers/__init__.py                               5      0   100%
/home/rmeli/git/molearn/src/molearn/trainers/openmm_physics_trainer.py                39     28    28%
/home/rmeli/git/molearn/src/molearn/trainers/sinkhorn_trainer.py                     206    170    17%
/home/rmeli/git/molearn/src/molearn/trainers/torch_physics_trainer.py                 38     27    29%
/home/rmeli/git/molearn/src/molearn/trainers/trainer.py                              202    169    16%
test_openmm_plugin.py                                                                138      1    99%
------------------------------------------------------------------------------------------------------
TOTAL                                                                               1996   1472    26%
```
