### 2023-07-10 Degiacomi-Lab/molearn#6 @SCMusson
I've had some issues with installing packages using pip. Primarily from version issues when installing beside packages from conda-forge. Additionally, having dependencies in  `setup.py` means dependencies are specified in two separate places (the other being the conda-forge feedstock).

I'm thinking a nice clean solution would be rely on conda-forge for dependencies completely. We could have it so that if you wish to install from source it is possible to get all the correct dependencies using:
`mamba install -c conda-forge --only-deps molearn`
and then install molearn manually with `setup.py`:
`python -m pip install . -v`

Do you think that would be an acceptable solution for you?
