### 2023-07-10 Degiacomi-Lab/molearn#6 @RMeli
@SCMusson yes, I think what you suggest is very sensible.
