### 2023-07-11 d1808cd
remove dependencies from setup.py
