### 2023-07-12 65fa95e
Update README.md with alternative installation methods
Modified installation methods.
