### 2023-07-14 Degiacomi-Lab/molearn#6 @RMeli
Thanks for updating the `README`. LGTM!

65fa95e
