### 2023-07-14 #5523 @RMeli
Thanks for all the improvements @degiacom! I never realized how much progress bars could improve the UX. I think the only outstanding issue is the NGLview problem discussed in https://github.com/Degiacomi-Lab/molearn_notebook/issues/3, otherwise LGTM!
