### 2023-07-17 #5523 @degiacom
Thank you @RMeli! We are still looking into the issue with NGLView in the notebooks. It appears it all boils down to  identifying and pinning a combination of `nglview`, `ipywidgets`, and `jupyter` package versions working correctly with each other.

While we chase that, @JoaoRodrigues, any chance you could provide us with your feedback too?
