### 2023-07-18 #5523 @degiacom
@RMeli, we finally seem to have identified the problem with NGLview, see [here](https://github.com/Degiacomi-Lab/molearn_notebook/issues/3#issuecomment-1640555591)!
