### 2023-07-18 a89084c
default for RMSD should be align, plus bug fix
