### 2023-07-18 #5523 @RMeli
@degiacom thank you for tracking down the issue! I can confirm that the suggested fix works for me.

@JoaoRodrigues the floor is yours! ;)
