### 2023-07-19 Degiacomi-Lab/molearn#10 @JoaoRodrigues
[JOSS] Manuscript Comments
Hello molearn team,

Please find some comments below regarding your manuscript submitted to JOSS.

## Statement of Need

On lines 21-22, I'd clarify that MD simulations offer an _approximation_ of atomistic dynamics of biomolecules and make a clear distinction to the data obtained from experiments. These are, after all, simulations. Training NNs on these dynamics is creating a model to reproduce the approximation, not "real" dynamics. You can mention it right at the end when you say that MD is not a "silver bullet".

From a purely formatting perspective, I'd add a line break before you introduce molearn, on line 40.

## Package Description

I'd remove the nested parenthesis on line 45-46 for the biobox reference.

On line 60, it should read "structure quality". On the same line, "root mean square deviation" should be explicity written of what. I assume atomic coordinates.

--

Other than these small comments, the manuscript reads very nicely and concisely.
