### 2023-07-19 Degiacomi-Lab/molearn#11 @JoaoRodrigues
[JOSS] General Code Comments
Hi molearn team,

Instead of leaving multiple small issues in the repository, I'll organize my comments here. If I have a larger comment or question, I'll open a separate issue and link here.

As a more general comment, I'd recomment using a tool like flake8 or if you're not in the mood to correct things manually, black/ruff, to standardize your code and have it comply a little more with Python style standards. Tiny things like spacing between operands, spacing between functions/docstrings, etc that make the code easier on the eyes for outsiders!

1. Regarding the `environment.yml` file, I am a little confused if it's meant to help people install from source or not. I see the instructions mention using `conda` or `mamba` with `-only-deps` and then running `pip install .`. Why not recommending users to simply run `conda env create -f environment.yml` as a one-step process?

2. Regarding the "no installation" instruction, I'd strongly recommend suggesting `pip install -e .`, which creates a symlink to the project but does not copy files to the Python installation folder.

3. I don't see a link to the documentation in the website linked to the project. Maybe it would be nice to include it there for simplicity and visibility? Since the docs seem to be mostly API docs, I'd rename them to molearn API docs though.

Assorted comments on code style:

1. You don't need the additional list inside the `str.join` [here](https://github.com/Degiacomi-Lab/molearn/blob/master/src/molearn/utils.py#L13). You can also use `_` as the variable for the iteration as an indication that you don't actually need it/use it.

2. Your `cpu_count` function [here](https://github.com/Degiacomi-Lab/molearn/blob/master/src/molearn/utils.py#L26) can be replaced by `multiprocessing.cpu_count` or `os.cpu_count` if I read it correctly. That should be more robust.

3. Python 3 classes do not need the `class Foo(object)` construction, they can be simply `class Foo`.

4.  You should guard these import errors with a variable (for instance `_has_modeller`) and check it whenever you call a function that calls the dependency. For instance [here](https://github.com/Degiacomi-Lab/molearn/blob/master/src/molearn/analysis/analyser.py#L17) there is nothing preventing the user from calling [this](https://github.com/Degiacomi-Lab/molearn/blob/master/src/molearn/analysis/analyser.py#L392) and crashing. So, the import check is a bit useless IMO. You could give the user a nicer message when they try to use the method that needs the dependency. Ideally, you'd move these import checks to `utils.py` or so. Another suggestion/idea is to move the analysers that depend on each third-party tool to separate files (e.g. `analyser_modeler.py`) and import those methods in the main `analyser.py` module. Just to separate the logic a little.

--

Life calls, will continue looking into the code tomorrow!
