+++ paper/paper.md
@@ -47,7 +47,7 @@ Here we present `molearn`, a Python framework facilitating the implementation of
 *	*Loss function definition*. While the classical loss function in a generative model typically builds upon a mean square error between input and output, here we provide the capability of directly interacting with the OpenMM molecular dynamics engine [@eastman2017openmm]. Specifically, we have implemented means of transferring PyTorch Tensor data directly into OpenMM's backend on GPU (without data transfer via the CPU). This enables quickly evaluating the energy of a generated model according to any force field accepted by OpenMM. This also enables directly running MD simulations with generated conformers while the model trains.
 *	*Model analysis*. Once a model is trained, it is important to gather metrics defining the quality of the protein structures it can generate. We provide tools to quickly quantify structure quality in terms of root mean square deviation between the coordinates of atoms in input and output, DOPE score [@shen2006statistical], Ramachandran plot, and user-defined functions. Analysis data is returned in the form of numpy arrays [@harris2020array], for ease of manipulation and plotting. We also provide a graphical user interface (GUI), enabling the visualisation of neural network latent space, and generation of interpolations between states visualised in an interactive 3D panel by a combination of MDAnalysis [@michaud2011mdanalysis] and nglview [@nguyen2018nglview].
 
-![`molearn` analysis tools include a graphical user interface, enabling the on-demand generation of protein conformations. The panel on the left controls how the neural network latent space is presented, the central panel is a Plotly interactive graph displaying the latent space, and the panel on the right is a 3D representation of an interpolation through the latent space supported by NGLview\label{fig:gui}](gui_MurD.png).
+![`molearn` analysis tools include a graphical user interface, enabling the on-demand generation of protein conformations. The panel on the left controls how the neural network latent space is presented, the central panel is a Plotly interactive graph displaying the latent space, and the panel on the right is a 3D representation of an interpolation through the latent space supported by NGLview\label{fig:gui}.](gui_MurD.png)
 
 
 # Usage
