### 2023-07-19 #5523 @degiacom
@JoaoRodrigues, following your [comments on the paper](https://github.com/Degiacomi-Lab/molearn/issues/10#issue-1811042482), above you will find an updated version (turns out, the formatting here slightly differs from the preview service on [Whedon](https://res.cloudinary.com/hkvhhndao/image/upload/v1689758839/zko9okninv6ydamatv57.pdf), which slightly complicates getting all the text to fit nicely with page breaks).
