### 2023-07-20 fafeab7
formatting and small cleanup
- first pass fixing indentations, spaces, and syntax details (guided by flake8)
- using os.cpu_count() instead of our custom function
