### 2023-07-20 Degiacomi-Lab/molearn#11 @degiacom
@JoaoRodrigues, thank you for your feedback. Here is a summary of our work in progress.

- did a first pass improving code formatting guided by flake8 (also removed some unused imports in the process)
- now all class declaration use Python 3 syntax
- we replaced our custom `utils.cpu_count` function with `os.cpu_count`. Thank you for pointing that out, much more robust!
- We had a reference to the documentation website at the bottom of the main Readme file, but it was not immediately findable. We now also refer to it at the top of the page, indicating that the documentation features "API and FAQ".

Better guarding the calls to methods requiring 3rd party software seems useful, we will look into that.
Concerning the `environment.yml` file, it is currently used by `.readthedocs.yaml`. This said, telling the user to exploit this file to create an environment does not seem like a bad idea to me. @SCMusson, what do you think?
