### 2023-07-21 Degiacomi-Lab/molearn#11 @SCMusson
Using `mamba install -c conda-forge --only-deps molearn` installs only the core dependencies and we have access to the conda-forge CI to confirm these are working (especially since we don't currently have our own CI setup yet). This is useful for a lean/minimum install and to avoid licensing issues (e.g. modeller licenses). We can add additional instructions using `conda env update --file environment.yml` or `conda env create -f environment.yml` that would also install all optional packages but this would then require the user to get a modeller license and either setup their environment variables or modify the config.py license key file for modeller manually. It is far far from a simple one step process unless salilab have changed how they are distributing modeller since the last time I installed it.
