### 2023-07-24 Degiacomi-Lab/molearn#11 @degiacom
I personally like the idea of having a lean installation, especially considering that in any case the Modeller manual license makes a fully automated installation impossible anyway. We have updated the *Installation* section of the README files to make instructions clearer, hopefully this will be ok.
