### 2023-07-28 Degiacomi-Lab/molearn#10 @degiacom
@JoaoRodrigues, just checking if you are happy with the changes. If so, happy to close this issue?
