### 2023-07-31 Degiacomi-Lab/molearn#11 @degiacom
@JoaoRodrigues, could you tell us if you are happy with the updates described above? If so, happy for the issue to be closed?
