### 2023-08-03 #5523 @degiacom
Hi @JoaoRodrigues, was wondering could you confirm whether you are happy with the way we addressed the issues you raised in the molearn repo, on [paper](https://github.com/Degiacomi-Lab/molearn/issues/10#issue-1811042482) and [code cleanliness](https://github.com/Degiacomi-Lab/molearn/issues/11#issue-1811068840)? If so, happy to close the issues?
