### 2023-08-04 379b238
GUI bugfix
prevent error when loading a previous GUI state without waypoint
