### 2023-08-11 #5523 @JoaoRodrigues
Hi Richard & Matteo,

Sorry for going MIA but I had some unexpected personal matters to attend to
and I couldn't get back to the review.

I had planned going over the code in a little more detail but I was pretty
happy with the paper and the code as it was. I would have only had minor
suggestions to make.

I saw the authors addressed my minor comments on the manuscript and some
others on the code/repo. I'm pretty happy with their work so feel free to
go ahead and greenlight this for publication.

Thank you again for the invitation and apologies for the delays!
