### 2023-08-25 f0254ff
Handle Lack of Iotbx/cctbx and modeller better. Error messages shoulfd be clearer
