### 2023-08-25 #5523 @SCMusson
@richardjgowers
I have created release v2.0.3 and set up Zenodo to give: https://doi.org/10.5281/zenodo.8284102

[![DOI](https://zenodo.org/badge/145391811.svg)](https://zenodo.org/badge/latestdoi/145391811)
