### 2023-08-25 #5523 @degiacom
Authors and affiliations in the article are correct
