### 2023-08-26 #5523 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
