### 2023-09-04 #5523 @Kevin-Mattheus-Moerman
@degiacom I am the AEiC for this track and here to help process final steps. I have check this review, the repository, the paper, and the archive, and most seems in order. I only have the below points that require your attention:

- [ ] in your affiliations please spell out `UK` as `United Kingdom`.
