### 2023-09-05 #5523 @Kevin-Mattheus-Moerman
@degiacom thanks. All looks good now. I did notice that although you have contributing guidelines in an CONTRIBUTING.md file, you do not link to it in your README. We can proceed here but I would recommend that you also have a Contributing heading in the README and link to that file.

I'll now proceed to process this for acceptance in JOSS.
