### 2023-09-05 #5523 @degiacom
Thank you for your feedback @Kevin-Mattheus-Moerman, I have edited the README file to point to CONTRIBUTING.md. Thank you also for accepting our work!
